Chore: Clean up code
