+++ code/gamelogic.h
@@ -5,3 +5,4 @@ short CheckPos(float X, float Y, bool isPlayer = false);
 bool isDestroyed(float x, float y);
 void attackTarget(int x, int y);
 void attackMobs(double x, double y);
+void handleMouseClick();
