+++ code/main.h
@@ -7,7 +7,6 @@ void generateRandomMap();
 void drawGame(HDC hdc);
 void handleInput();
 void updateMouseHover();
-void handleMouseClick();
 void SummonMobs();
 void ToggleBackpackWindow(HWND parentHwnd, HINSTANCE hInstance);
 bool isPathClear(float startX, float startY, float endX, float endY);
